requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.1.0
backcall==0.2.0
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==1.0.0
Django==4.0.6
django-braces==1.15.0
django-extensions==3.2.1
django-localflavor==3.1
django-model-utils==4.2.0
django-oauth-toolkit==2.2.0
django-rest-framework-social-oauth2==1.1.0
django-seed==0.3.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
executing==1.2.0
Faker==15.1.3
flake8==5.0.4
idna==3.4
ipdb==0.13.9
ipython==8.6.0
isort==5.10.1
jedi==0.18.1
jwcrypto==1.4.2
matplotlib-inline==0.1.6
mccabe==0.7.0
oauthlib==3.2.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
Pygments==2.13.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
python-stdnum==1.17
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
stack-data==0.6.0
toml==0.10.2
toposort==1.7
traitlets==5.5.0
urllib3==1.26.12
wcwidth==0.2.5
wrapt==1.14.1
